# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TEST",
        "title": "API",
        "contact": {},
        "version": "1.7"
    },
    "basePath": "/api/v1",
    "paths": {
        "/athletes/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint adds a new athlete.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Athlete"
                ],
                "summary": "Add an athlete",
                "parameters": [
                    {
                        "description": "Athlete details to add",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Athlete"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Athlete"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Message"
                        }
                    }
                }
            }
        },
        "/athletes/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint deletes an athlete by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Athlete"
                ],
                "summary": "Delete an athlete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Athlete ID to delete",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Message"
                        }
                    }
                }
            }
        },
        "/athletes/edit": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint edits an existing athlete.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Athlete"
                ],
                "summary": "Edit an athlete",
                "parameters": [
                    {
                        "description": "Athlete details to edit",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Athlete"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Athlete"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Message"
                        }
                    }
                }
            }
        },
        "/athletes/get": {
            "get": {
                "description": "This endpoint retrieves an athlete by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Athlete"
                ],
                "summary": "Get an athlete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Athlete ID to get",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Athlete"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Message"
                        }
                    }
                }
            }
        },
        "/athletes/getall": {
            "get": {
                "description": "This endpoint retrieves all athletes with pagination and optional filters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Athlete"
                ],
                "summary": "List athletes",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Country ID filter",
                        "name": "country_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sport type filter",
                        "name": "sport_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.ListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "olympy_api-gateway_genproto_athlete_service.Athlete": {
            "type": "object",
            "properties": {
                "country_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sport_type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "olympy_api-gateway_genproto_athlete_service.ListResponse": {
            "type": "object",
            "properties": {
                "athletes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/olympy_api-gateway_genproto_athlete_service.Athlete"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "olympy_api-gateway_genproto_athlete_service.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}